Fix FileNotFoundException

+++ tools/CustomBuildTool/HeaderGen.cs
@@ -232,11 +232,15 @@ namespace CustomBuildTool
             {
                 string headerFileName = Path.Join([BaseDirectory, OutputFile]);
                 string headerUpdateText = sw.ToString();
-                string headerCurrentText = Utils.ReadAllText(headerFileName);
 
-                if (!string.Equals(headerUpdateText, headerCurrentText, StringComparison.OrdinalIgnoreCase))
+                if (File.Exists(headerFileName))
                 {
-                    Utils.WriteAllText(headerFileName, headerUpdateText);
+                    string headerCurrentText = Utils.ReadAllText(headerFileName);
+
+                    if (!string.Equals(headerUpdateText, headerCurrentText, StringComparison.OrdinalIgnoreCase))
+                    {
+                        Utils.WriteAllText(headerFileName, headerUpdateText);
+                    }
                 }
 
                 Program.PrintColorMessage($"HeaderGen -> {headerFileName}", ConsoleColor.Cyan);
